# openapi.yaml
openapi: 3.1.0
info:
  title: Last Slot API
  version: 1.0.0
  description: Public contract for the Last Slot reliability case study.
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        '200':
          description: Gateway is ready.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /v1/slots/{slotId}:
    get:
      operationId: getSlot
      parameters:
        - $ref: '#/components/parameters/SlotId'
      responses:
        '200':
          description: Current slot state.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slot'
        '404':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/slots/{slotId}/bookings:
    post:
      operationId: createBooking
      parameters:
        - $ref: '#/components/parameters/SlotId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [customerName]
              properties:
                customerName:
                  type: string
                  minLength: 2
                  maxLength: 80
      responses:
        '201':
          description: >-
            Booking created. There is no separate booking resource; the
            persisted state is read back through GET /v1/slots/{slotId}.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResult'
        '200':
          description: Idempotent replay of the original booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResult'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '415':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Unavailable'
components:
  parameters:
    SlotId:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Booking:
      type: object
      additionalProperties: false
      required: [id, customerName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        customerName:
          type: string
        createdAt:
          type: string
          format: date-time
    Slot:
      type: object
      additionalProperties: false
      required: [id, title, startsAt, status]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        startsAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [available, booked]
        booking:
          oneOf:
            - $ref: '#/components/schemas/Booking'
            - type: 'null'
    BookingResult:
      type: object
      additionalProperties: false
      required: [booking, slot, replayed]
      properties:
        booking:
          $ref: '#/components/schemas/Booking'
        slot:
          $ref: '#/components/schemas/Slot'
        replayed:
          type: boolean
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
  responses:
    Error:
      description: Canonical error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: A dependency is temporarily unavailable.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'